Fixed windows compilation. Works now well on Linux and Windows Fixed the z test for HiZ. Still completely messy code to clean up but it works well. Frustum culling is wrong and should be fixed soon.

+++ src/rt/bvh2.hpp
@@ -53,8 +53,8 @@ namespace pf
     INLINE BVH2BuildOption(void) {}
     INLINE BVH2BuildOption(uint32 minPrimNum,
                            uint32 maxPrimNum,
-                           uint32 SAHIntersectionCost,
-                           uint32 SAHTraversalCost) :
+                           float SAHIntersectionCost,
+                           float SAHTraversalCost) :
       minPrimNum(minPrimNum),
       maxPrimNum(maxPrimNum),
       SAHIntersectionCost(SAHIntersectionCost),
